third_party/WebKit/LayoutTests/permissionclient/audio-permissions-expected.txt

MockContentSettingsClient: allowMedia((file test):media/content/silence.oga): true
MockContentSettingsClient: allowMedia((file test):media/content/silence.oga?nocache): false
MockContentSettingsClient: allowMedia((file test):media/content/silence.oga): false
MockContentSettingsClient: allowMedia((file test):media/content/silence.oga): false

PASS: first sound loaded
